@@ -62,54 +62,52 @@
     }
   },
   "devDependencies": {
-    "@aws-sdk/credential-providers": "^3.515.0",
+    "@aws-sdk/credential-providers": "^3.632.0",
     "@iarna/toml": "^2.2.5",
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
-    "@microsoft/api-extractor": "^7.43.1",
-    "@microsoft/tsdoc-config": "^0.16.2",
+    "@microsoft/api-extractor": "^7.47.5",
+    "@microsoft/tsdoc-config": "^0.17.0",
     "@mongodb-js/zstd": "^1.2.0",
-    "@types/chai": "^4.3.14",
+    "@types/chai": "^4.3.17",
     "@types/chai-subset": "^1.3.5",
     "@types/express": "^4.17.21",
     "@types/kerberos": "^1.1.5",
-    "@types/mocha": "^10.0.6",
-    "@types/node": "^20.14.10",
+    "@types/mocha": "^10.0.7",
+    "@types/node": "^22.4.1",
     "@types/saslprep": "^1.0.3",
     "@types/semver": "^7.5.8",
     "@types/sinon": "^17.0.3",
     "@types/sinon-chai": "^3.2.12",
-    "@types/whatwg-url": "^11.0.4",
-    "@typescript-eslint/eslint-plugin": "^5.62.0",
-    "@typescript-eslint/parser": "^5.62.0",
+    "@types/whatwg-url": "^11.0.5",
+    "@typescript-eslint/eslint-plugin": "^8.2.0",
+    "@typescript-eslint/parser": "^8.2.0",
     "chai": "^4.4.1",
     "chai-subset": "^1.6.0",
     "chalk": "^4.1.2",
-    "eslint": "^8.56.0",
-    "eslint-config-prettier": "^8.10.0",
-    "eslint-plugin-github": "^4.10.2",
-    "eslint-plugin-import": "^2.29.1",
+    "eslint": "^9.0.0",
+    "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-mocha": "^10.4.1",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-simple-import-sort": "^10.0.0",
-    "eslint-plugin-tsdoc": "^0.2.17",
-    "eslint-plugin-unused-imports": "^2.0.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-simple-import-sort": "^12.1.1",
+    "eslint-plugin-tsdoc": "^0.3.0",
+    "eslint-plugin-unused-imports": "^4.0.0",
     "express": "^4.19.2",
     "gcp-metadata": "^5.3.0",
     "js-yaml": "^4.1.0",
     "mocha": "^10.4.0",
     "mocha-sinon": "^2.1.2",
     "mongodb-client-encryption": "^6.1.0",
-    "mongodb-legacy": "^6.0.1",
+    "mongodb-legacy": "^6.1.0",
     "nyc": "^15.1.0",
-    "prettier": "^2.8.8",
-    "semver": "^7.6.0",
-    "sinon": "^17.0.1",
+    "prettier": "^3.3.3",
+    "semver": "^7.6.3",
+    "sinon": "^18.0.0",
     "sinon-chai": "^3.7.0",
     "snappy": "^7.2.2",
     "socks": "^2.8.1",
     "source-map-support": "^0.5.21",
     "ts-node": "^10.9.2",
-    "tsd": "^0.31.0",
+    "tsd": "^0.31.1",
     "typescript": "5.5",
     "typescript-cached-transpile": "^0.0.6",
     "v8-heapsnapshot": "^1.3.1",
