@@ -1,4 +1,4 @@
-import path = require('path');
+import * as path from 'path';
 
 import { loadSpecTests } from '../../spec';
 import { runUnifiedSuite } from '../../tools/unified-spec-runner/runner';
