@@ -1,6 +1,6 @@
 import { promisify } from 'util';
 
-import { type BSONSerializeOptions, type Document, resolveBSONOptions } from '../bson';
+import { type BSONSerializeOptions, type Document, EJSON, resolveBSONOptions } from '../bson';
 import type { Collection } from '../collection';
 import {
   type AnyError,
@@ -20,12 +20,12 @@ import { makeUpdateStatement, UpdateOperation, type UpdateStatement } from '../o
 import type { Server } from '../sdam/server';
 import type { Topology } from '../sdam/topology';
 import type { ClientSession } from '../sessions';
-import { maybeAddIdToDocuments } from '../utils';
 import {
   applyRetryableWrites,
   type Callback,
   getTopology,
   hasAtomicOperators,
+  maybeAddIdToDocuments,
   type MongoDBNamespace,
   resolveOptions
 } from '../utils';
@@ -294,7 +294,7 @@ export class BulkWriteResult {
   }
 
   toString(): string {
-    return `BulkWriteResult(${this.result})`;
+    return `BulkWriteResult(${EJSON.stringify(this.result)})`;
   }
 
   isOk(): boolean {
@@ -583,13 +583,12 @@ function executeCommands(
     const operation = isInsertBatch(batch)
       ? new InsertOperation(bulkOperation.s.namespace, batch.operations, finalOptions)
       : isUpdateBatch(batch)
-      ? new UpdateOperation(bulkOperation.s.namespace, batch.operations, finalOptions)
-      : isDeleteBatch(batch)
-      ? new DeleteOperation(bulkOperation.s.namespace, batch.operations, finalOptions)
-      : null;
+        ? new UpdateOperation(bulkOperation.s.namespace, batch.operations, finalOptions)
+        : isDeleteBatch(batch)
+          ? new DeleteOperation(bulkOperation.s.namespace, batch.operations, finalOptions)
+          : null;
 
     if (operation != null) {
-      // eslint-disable-next-line github/no-then
       executeOperation(bulkOperation.s.collection.client, operation).then(
         result => resultHandler(undefined, result),
         error => resultHandler(error)
@@ -919,7 +918,11 @@ export abstract class BulkOperationBase {
    * Create a new OrderedBulkOperation or UnorderedBulkOperation instance
    * @internal
    */
-  constructor(private collection: Collection, options: BulkWriteOptions, isOrdered: boolean) {
+  constructor(
+    private collection: Collection,
+    options: BulkWriteOptions,
+    isOrdered: boolean
+  ) {
     // determine whether bulkOperation is ordered or unordered
     this.isOrdered = isOrdered;
 
