@@ -49,7 +49,7 @@ export class MongocryptdManager {
   async spawn(): Promise<void> {
     const cmdName = this.spawnPath || 'mongocryptd';
 
-    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    // eslint-disable-next-line @typescript-eslint/no-require-imports
     const { spawn } = require('child_process') as typeof import('child_process');
 
     // Spawned with stdio: ignore and detached: true
