@@ -1,7 +1,6 @@
 import { Writable } from 'stream';
 
-import type { Document } from '../bson';
-import { ObjectId } from '../bson';
+import { type Document, ObjectId } from '../bson';
 import type { Collection } from '../collection';
 import { MongoAPIError, MONGODB_ERROR_CODES, MongoError } from '../error';
 import { type Callback, squashError } from '../utils';
@@ -135,7 +134,6 @@ export class GridFSBucketWriteStream extends Writable {
     if (!this.bucket.s.calledOpenUploadStream) {
       this.bucket.s.calledOpenUploadStream = true;
 
-      // eslint-disable-next-line github/no-then
       checkIndexes(this).then(() => {
         this.bucket.s.checkedIndexes = true;
         this.bucket.emit('index');
@@ -272,7 +270,6 @@ function checkDone(stream: GridFSBucketWriteStream, callback: Callback): void {
       return;
     }
 
-    // eslint-disable-next-line github/no-then
     stream.files.insertOne(gridFSFile, { writeConcern: stream.writeConcern }).then(
       () => {
         stream.gridFSFile = gridFSFile;
@@ -396,7 +393,6 @@ function doWrite(
         return;
       }
 
-      // eslint-disable-next-line github/no-then
       stream.chunks.insertOne(doc, { writeConcern: stream.writeConcern }).then(
         () => {
           --stream.state.outstandingRequests;
@@ -437,7 +433,6 @@ function writeRemnant(stream: GridFSBucketWriteStream, callback: Callback): void
     return;
   }
 
-  // eslint-disable-next-line github/no-then
   stream.chunks.insertOne(doc, { writeConcern: stream.writeConcern }).then(
     () => {
       --stream.state.outstandingRequests;
