@@ -209,7 +209,7 @@ export class FindAndModifyOperation extends CommandOperation<Document> {
 
     // Execute the command
     const result = await super.executeCommand(server, session, cmd);
-    return options.includeResultMetadata ? result : result.value ?? null;
+    return options.includeResultMetadata ? result : (result.value ?? null);
   }
 }
 
