@@ -17,11 +17,6 @@ export const Aspect = {
 /** @public */
 export type Hint = string | Document;
 
-// eslint-disable-next-line @typescript-eslint/ban-types
-export interface OperationConstructor extends Function {
-  aspects?: Set<symbol>;
-}
-
 /** @public */
 export interface OperationOptions extends BSONSerializeOptions {
   /** Specify ClientSession for this command */
@@ -63,10 +58,12 @@ export abstract class AbstractOperation<TResult = any> {
 
   [kSession]: ClientSession | undefined;
 
+  static aspects?: Set<symbol>;
+
   constructor(options: OperationOptions = {}) {
     this.readPreference = this.hasAspect(Aspect.WRITE_OPERATION)
       ? ReadPreference.primary
-      : ReadPreference.fromOptions(options) ?? ReadPreference.primary;
+      : (ReadPreference.fromOptions(options) ?? ReadPreference.primary);
 
     // Pull the BSON serialize options from the already-resolved options
     this.bsonOptions = resolveBSONOptions(options);
@@ -85,7 +82,7 @@ export abstract class AbstractOperation<TResult = any> {
   abstract execute(server: Server, session: ClientSession | undefined): Promise<TResult>;
 
   hasAspect(aspect: symbol): boolean {
-    const ctor = this.constructor as OperationConstructor;
+    const ctor = this.constructor as { aspects?: Set<symbol> };
     if (ctor.aspects == null) {
       return false;
     }
@@ -111,7 +108,7 @@ export abstract class AbstractOperation<TResult = any> {
 }
 
 export function defineAspects(
-  operation: OperationConstructor,
+  operation: { aspects?: Set<symbol> },
   aspects: symbol | symbol[] | Set<symbol>
 ): Set<symbol> {
   if (!Array.isArray(aspects) && !(aspects instanceof Set)) {
