@@ -95,7 +95,6 @@ export class SrvPoller extends TypedEventEmitter<SrvPollerEvents> {
     }
 
     this._timeout = setTimeout(() => {
-      // eslint-disable-next-line github/no-then
       this._poll().then(undefined, squashError);
     }, this.intervalMS);
   }
@@ -117,7 +116,7 @@ export class SrvPoller extends TypedEventEmitter<SrvPollerEvents> {
 
     try {
       srvRecords = await dns.promises.resolveSrv(this.srvAddress);
-    } catch (dnsError) {
+    } catch {
       this.failure();
       return;
     }
