@@ -2,8 +2,7 @@ import type { Document } from './bson';
 import { MongoRuntimeError, MongoTransactionError } from './error';
 import type { CommandOperationOptions } from './operations/command';
 import { ReadConcern, type ReadConcernLike } from './read_concern';
-import type { ReadPreferenceLike } from './read_preference';
-import { ReadPreference } from './read_preference';
+import { ReadPreference, type ReadPreferenceLike } from './read_preference';
 import type { Server } from './sdam/server';
 import { WriteConcern } from './write_concern';
 
