@@ -1,24 +1,28 @@
-
+import { expect } from 'chai';
 import { describe, it } from 'mocha';
-import { AbstractCursor, ChangeStream, ClientSession, GridFSBucket, MongoClient } from 'mongodb/lib/beta';
+import {
+  AbstractCursor,
+  ChangeStream,
+  ClientSession,
+  GridFSBucket,
+  MongoClient
+} from 'mongodb/lib/beta';
+import * as sinon from 'sinon';
 import { Readable } from 'stream';
 import { pipeline } from 'stream/promises';
-import { expect } from 'chai';
 import { setTimeout } from 'timers/promises';
-import { createReadStream } from 'fs';
-import { join } from 'path';
-
-import * as sinon from 'sinon';
 
 async function setUpCollection(client: MongoClient) {
   const collection = client.db('foo').collection<{ name: string }>('bar');
   const documents: Array<{ name: string }> = Array.from({ length: 5 }).map(i => ({
     name: String(i)
   }));
-  await collection.insertMany(documents)
+  await collection.insertMany(documents);
   return collection;
 }
 
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+const MONGODB_URI = process.env.MONGODB_URI!;
 describe('explicit resource management feature integration tests', function () {
   const clientDisposeSpy = sinon.spy(MongoClient.prototype, Symbol.asyncDispose);
   const sessionDisposeSpy = sinon.spy(ClientSession.prototype, Symbol.asyncDispose);
@@ -26,27 +30,28 @@ describe('explicit resource management feature integration tests', function () {
   const cursorDisposeSpy = sinon.spy(AbstractCursor.prototype, Symbol.asyncDispose);
   const readableDisposeSpy = sinon.spy(Readable.prototype, Symbol.asyncDispose);
 
-  afterEach(function(){
+  afterEach(function () {
     clientDisposeSpy.resetHistory();
     sessionDisposeSpy.resetHistory();
     changeStreamDisposeSpy.resetHistory();
     cursorDisposeSpy.resetHistory();
     readableDisposeSpy.resetHistory();
-  })
+  });
+
   describe('MongoClient', function () {
     it('does not crash or error when used with await-using syntax', async function () {
       {
-        await using client = new MongoClient(process.env.MONGODB_URI!);
+        await using client = new MongoClient(MONGODB_URI);
         await client.connect();
       }
       expect(clientDisposeSpy.called).to.be.true;
-    })
-  })
+    });
+  });
 
   describe('Cursors', function () {
     it('does not crash or error when used with await-using syntax', async function () {
       {
-        await using client = new MongoClient(process.env.MONGODB_URI!);
+        await using client = new MongoClient(MONGODB_URI);
         await client.connect();
 
         const collection = await setUpCollection(client);
@@ -55,39 +60,39 @@ describe('explicit resource management feature integration tests', function () {
         await cursor.next();
       }
       expect(cursorDisposeSpy.called).to.be.true;
-    })
+    });
 
     describe('cursor streams', function () {
       it('does not crash or error when used with await-using syntax', async function () {
         {
-          await using client = new MongoClient(process.env.MONGODB_URI!);
+          await using client = new MongoClient(MONGODB_URI);
           await client.connect();
 
           const collection = await setUpCollection(client);
 
-          await using readable = collection.find().stream();
+          await using _readable = collection.find().stream();
         }
         expect(readableDisposeSpy.called).to.be.true;
-      })
-    })
-  })
+      });
+    });
+  });
 
   describe('Sessions', function () {
     it('does not crash or error when used with await-using syntax', async function () {
       {
-        await using client = new MongoClient(process.env.MONGODB_URI!);
+        await using client = new MongoClient(MONGODB_URI);
         await client.connect();
 
-        await using session = client.startSession();
+        await using _session = client.startSession();
       }
       expect(sessionDisposeSpy.called).to.be.true;
-    })
-  })
+    });
+  });
 
   describe('ChangeStreams', function () {
     it('does not crash or error when used with await-using syntax', async function () {
       {
-        await using client = new MongoClient(process.env.MONGODB_URI!);
+        await using client = new MongoClient(MONGODB_URI);
         await client.connect();
 
         const collection = await setUpCollection(client);
@@ -97,23 +102,23 @@ describe('explicit resource management feature integration tests', function () {
         await cs.next();
       }
       expect(changeStreamDisposeSpy.called).to.be.true;
-    })
+    });
   });
 
   describe('GridFSDownloadStream', function () {
     it('does not crash or error when used with await-using syntax', async function () {
       {
-        await using client = new MongoClient(process.env.MONGODB_URI!);
+        await using client = new MongoClient(MONGODB_URI);
         await client.connect();
 
         const bucket = new GridFSBucket(client.db('foo'));
-        const uploadStream = bucket.openUploadStream('foo.txt')
-        await pipeline(Readable.from("AAAAAAA".split('')), uploadStream);
+        const uploadStream = bucket.openUploadStream('foo.txt');
+        await pipeline(Readable.from('AAAAAAA'.split('')), uploadStream);
 
-        await using downloadStream = bucket.openDownloadStreamByName('foo.txt');
+        await using _downloadStream = bucket.openDownloadStreamByName('foo.txt');
       }
 
       expect(readableDisposeSpy.called).to.be.true;
-    })
+    });
   });
-})
+});
