@@ -62,10 +62,10 @@ describe('CRUD unified', function () {
       return description.match(clientBulkWriteTests)
         ? 'TODO(NODE-6257): implement client level bulk write'
         : unacknowledgedHintTests.includes(description)
-        ? `TODO(NODE-3541)`
-        : isLoadBalanced && loadBalancedCollationTests.includes(description)
-        ? `TODO(NODE-6280): fix collation for find and modify commands on load balanced mode`
-        : false;
+          ? `TODO(NODE-3541)`
+          : isLoadBalanced && loadBalancedCollationTests.includes(description)
+            ? `TODO(NODE-6280): fix collation for find and modify commands on load balanced mode`
+            : false;
     }
   );
 });
