@@ -5,7 +5,8 @@ import * as sinon from 'sinon';
 import { MongoClient } from '../mongodb';
 
 const expect = chai.expect;
-// eslint-disable-next-line @typescript-eslint/no-var-requires
+
+// eslint-disable-next-line @typescript-eslint/no-require-imports
 chai.use(require('sinon-chai'));
 
 async function verifyKerberosAuthentication(client) {
