@@ -1,7 +1,7 @@
 let snappy;
 try {
   snappy = require('snappy');
-} catch (error) {
+} catch {
   snappy = {
     uncompressSync(data) {
       console.warn('fake snappy uncompressSync call');
