@@ -1,7 +1,6 @@
 /* eslint-disable simple-import-sort/imports */
-/* eslint-disable import/first */
 
-// eslint-disable-next-line @typescript-eslint/no-var-requires
+// eslint-disable-next-line @typescript-eslint/no-require-imports
 require('source-map-support').install({
   hookRequire: true
 });
@@ -200,8 +199,10 @@ const cleanUpMocksAfterHook = () => mock.cleanup();
 
 const beforeAllPluginImports = () => {
   // optionally enable test runner-wide plugins
+  // eslint-disable-next-line @typescript-eslint/no-require-imports
   require('../plugins/deferred');
   // configure mocha
+  // eslint-disable-next-line @typescript-eslint/no-require-imports
   require('mocha-sinon');
 };
 
