@@ -1,8 +1,7 @@
 import { once } from 'events';
 import { expectType } from 'tsd';
 
-import type { ConnectionPoolCreatedEvent } from '../mongodb';
-import { MongoClient } from '../mongodb';
+import { type ConnectionPoolCreatedEvent, MongoClient } from '../mongodb';
 
 const client: MongoClient = new MongoClient('');
 const p = once(client, 'connectionPoolCreated');
