@@ -1,7 +1,6 @@
 import { expectNotType, expectType } from 'tsd';
 
-import type { CollectionInfo, ListCollectionsCursor } from '../mongodb';
-import { MongoClient } from '../mongodb';
+import { type CollectionInfo, type ListCollectionsCursor, MongoClient } from '../mongodb';
 
 const db = new MongoClient('').db();
 
