@@ -1,7 +1,6 @@
 import { expectAssignable, expectError, expectNotAssignable, expectNotType, expectType } from 'tsd';
 
-import type { Collection, Document, WithId } from '../mongodb';
-import { ObjectId } from '../mongodb';
+import { type Collection, type Document, ObjectId, type WithId } from '../mongodb';
 
 type InsertOneFirstParam<Schema extends Document> = Parameters<Collection<Schema>['insertOne']>[0];
 
