@@ -16,8 +16,9 @@ export type PropExists<Type, Key extends string> = Key extends keyof Type ? true
 export type Expand<T> = T extends infer O ? { [K in keyof O]: O[K] } : never;
 
 // expands object types recursively
-export type ExpandRecursively<T> = T extends Record<string, unknown>
-  ? T extends infer O
-    ? { [K in keyof O]: ExpandRecursively<O[K]> }
-    : never
-  : T;
+export type ExpandRecursively<T> =
+  T extends Record<string, unknown>
+    ? T extends infer O
+      ? { [K in keyof O]: ExpandRecursively<O[K]> }
+      : never
+    : T;
