@@ -8,20 +8,18 @@ import { setTimeout } from 'timers';
 import { setTimeout as setTimeoutPromise } from 'timers/promises';
 
 import {
-  Long,
-  MongoClient,
-  ObjectId,
-  RTTSampler,
-  ServerHeartbeatSucceededEvent
-} from '../../mongodb';
-import {
   isHello,
   LEGACY_HELLO_COMMAND,
+  Long,
+  MongoClient,
   Monitor,
   MonitorInterval,
+  ObjectId,
+  RTTSampler,
   ServerDescription,
   type ServerHeartbeatFailedEvent,
   type ServerHeartbeatStartedEvent,
+  ServerHeartbeatSucceededEvent,
   ServerType
 } from '../../mongodb';
 import * as mock from '../../tools/mongodb-mock/index';
