@@ -48,9 +48,8 @@ describe('driver utils', function () {
 
         context('when the host matches a FQDN with subdomain', function () {
           it('returns true', function () {
-            expect(
-              hostMatchesWildcards('prod.mongodb.net', ['prod.mongodb.net', 'other'])
-            ).to.be.true;
+            expect(hostMatchesWildcards('prod.mongodb.net', ['prod.mongodb.net', 'other'])).to.be
+              .true;
           });
         });
 
@@ -96,9 +95,8 @@ describe('driver utils', function () {
 
         context('when the host matches a FQDN with subdomain', function () {
           it('returns true', function () {
-            expect(
-              hostMatchesWildcards('prod.mongodb.net', ['*.prod.mongodb.net', 'other'])
-            ).to.be.true;
+            expect(hostMatchesWildcards('prod.mongodb.net', ['*.prod.mongodb.net', 'other'])).to.be
+              .true;
           });
         });
 
@@ -145,17 +143,15 @@ describe('driver utils', function () {
     context('when using unix domain sockets', function () {
       context('when the host matches at least one', function () {
         it('returns true', function () {
-          expect(
-            hostMatchesWildcards('/tmp/mongodb-27017.sock', ['*/mongodb-27017.sock', 'other'])
-          ).to.be.true;
+          expect(hostMatchesWildcards('/tmp/mongodb-27017.sock', ['*/mongodb-27017.sock', 'other']))
+            .to.be.true;
         });
       });
 
       context('when the host does not match any', function () {
         it('returns false', function () {
-          expect(
-            hostMatchesWildcards('/tmp/mongodb-27017.sock', ['*/mongod-27017.sock', 'test2'])
-          ).to.be.false;
+          expect(hostMatchesWildcards('/tmp/mongodb-27017.sock', ['*/mongod-27017.sock', 'test2']))
+            .to.be.false;
         });
       });
     });
